Refactor agent tool code to make them modular
Assemble tools description in prompt from tools themselves as well
