+++ app/components/agent/logic/tools/tools.ts
@@ -1,0 +1,11 @@
+import { analyzeDocuments } from './analyzeDocs';
+import { createSuggestedPipeline } from './createSuggestedPipeline';
+import { refinePipeline } from './refinePipeline';
+import { suggestPipeline } from './suggestPipeline';
+
+export const agentTools = [
+  analyzeDocuments,
+  suggestPipeline,
+  refinePipeline,
+  createSuggestedPipeline,
+];
